+++ packages/css-packs/reduced-ui-pack/src/internal/iconIds.js
@@ -371,6 +371,7 @@ const iconIds: Array<string> = [
   'presence-unavailable',
   'question-circle',
   'question',
+  'questions',
   'queues',
   'quote',
   'radio',