+++ packages/editor/editor-core/__tests__/integration/_helpers.ts
@@ -4,6 +4,7 @@
  * And, don't get too fancy with it ;)
  */
 import { getExampleUrl } from '@atlaskit/webdriver-runner/utils/example';
+import { sleep } from '@atlaskit/editor-test-helpers';
 export const getDocFromElement = el => el.pmViewDesc.node.toJSON();
 export const editable = '.ProseMirror';
 
@@ -25,6 +26,8 @@ export const message = {
   placeholder: '.ProseMirror',
 };
 
+export const editors = [comment, fullpage];
+
 export const clipboardHelper = getExampleUrl(
   'editor',
   'editor-core',
@@ -35,6 +38,8 @@ export const clipboardInput = '#input';
 
 export const copyAsPlaintextButton = '#copy-as-plaintext';
 export const copyAsHTMLButton = '#copy-as-html';
+
+export const mediaInsertDelay = 500;
 
 const mediaPickerMock = '.mediaPickerMock';
 export const setupMediaMocksProviders = async browser => {
@@ -50,7 +55,7 @@ export const setupMediaMocksProviders = async browser => {
   await browser.click('.reloadEditorButton');
 };
 
-export const insertMedia = async (browser, indexes = [0]) => {
+export const insertMedia = async (browser, indexes = [-1]) => {
   const openMediaPopup = '[aria-label="Insert files and images"]';
   const insertMediaButton = '.e2e-insert-button';
 
@@ -68,4 +73,10 @@ export const insertMedia = async (browser, indexes = [0]) => {
   // insert it from the picker dialog
   await browser.waitForSelector(insertMediaButton);
   await browser.click(insertMediaButton);
+
+  // after clicking Insert media, the media plugin needs to upload the file,
+  // and generate some intermediate processing and preview states while that's happening
+  //
+  // we currently don't map the media state as a CSS selector
+  await sleep(mediaInsertDelay);
 };