+++ packages/elements/util-service-support/src/serviceUtils.ts
@@ -104,26 +104,24 @@ export const requestService = <T>(
     credentials,
   };
 
-  return fetch(new Request(requestUrl, requestOptions)).then(
-    (response: Response) => {
-      if (response.status === 204) {
-        return Promise.resolve();
-      } else if (response.ok) {
-        return response.json();
-      } else if (response.status === 401 && refreshedSecurityProvider) {
-        // auth issue - try once
-        return refreshedSecurityProvider().then(newSecOptions => {
-          const retryServiceConfig = {
-            url,
-            securityProvider: () => newSecOptions,
-          };
-          return requestService(retryServiceConfig, options);
-        });
-      }
-      return Promise.reject({
-        code: response.status,
-        reason: response.statusText,
+  return fetch(requestUrl, requestOptions).then((response: Response) => {
+    if (response.status === 204) {
+      return Promise.resolve();
+    } else if (response.ok) {
+      return response.json();
+    } else if (response.status === 401 && refreshedSecurityProvider) {
+      // auth issue - try once
+      return refreshedSecurityProvider().then(newSecOptions => {
+        const retryServiceConfig = {
+          url,
+          securityProvider: () => newSecOptions,
+        };
+        return requestService(retryServiceConfig, options);
       });
-    },
-  );
+    }
+    return Promise.reject({
+      code: response.status,
+      reason: response.statusText,
+    });
+  });
 };