+++ packages/editor/editor-core/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/editor-core
 
+## 79.0.4
+- [patch] ED-4680 add smart card plugin, enable for inline smart cards [b9529e6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b9529e6)
+
 ## 79.0.3
 - [patch] propagate sessionId to the mentionTypeahead rendered event and service endpoints [0c37666](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/0c37666)